+++ CHANGELOG.md
@@ -579,3 +579,4 @@
 * Added campaign start date and last save date to save metadata
 * Added OnCoreShutdown call to CCModule that gets called before any modules are disposed
 * Added BeforeCoreShutdown script hook called by OnCoreShutdown before any modules are disposed
+* Added TreatUnknownAsNull option to PxEnumConverter